Moved some prerelease commands to common
in preparation for sync subcommand
